def APNsetup(df): function set up (kinda works)
It skips over majority of the APNs, in the list but some it doesn't and I also get a warning I need to look in to.  But a better spot than I was in. Also after is runs I can no longer set APN as index.

SettingWithCopyWarning: 
A value is trying to be set on a copy of a slice from a DataFrame.
Try using .loc[row_indexer,col_indexer] = value instead